Consolidate game scene and phase data
