--- GameScene.js
@@ -1,637 +1,0 @@
-import Phaser from 'phaser';
-import DifficultyManager from '../utils/DifficultyManager';
-import { getHUDConfig } from '../utils/hudConfig';
-
-import pauseIcon from '../assets/images/ui/pause.png';
-import lifeFullIcon from '../assets/images/ui/life_full.png';
-import lifeEmptyIcon from '../assets/images/ui/life_empty.png';
-import scoreIcon from '../assets/images/ui/score.png';
-import timeIcon from '../assets/images/ui/time.png';
-import tipBgIcon from '../assets/images/ui/card_tip_bg.png';
-
-export default class GameScene extends Phaser.Scene {
-  constructor({ phaseConfig, bitmask, onGameOver, engine }) {
-    super({ key: 'GameScene' });
-    this.phaseConfig = phaseConfig;
-    this.bitmask = bitmask;
-    this.onGameOver = onGameOver;
-    this.engine = engine;
-    this.score = 0;
-    this.lives = 3;
-    this.activeItems = [];
-    this.lastSpawn = 0;
-    this.difficulty = new DifficultyManager();
-    this.baseSpawnRate = phaseConfig.itemSpawnRate || phaseConfig.spawnRate || 1500;
-    this.baseSimultaneous = phaseConfig.simultaneous || 2;
-    this.spawnRate = this.baseSpawnRate;
-    this.simultaneous = this.baseSimultaneous;
-    this.speed = phaseConfig.speed || 1;
-    this.duration = phaseConfig.duration || null;
-    this.tips = phaseConfig.tips || [];
-    this.music = phaseConfig.music || '/assets/audio/background.ogg';
-    this.tipText = null;
-    this.tipCard = null;
-    this.lifeIcons = [];
-    this.chef = null;
-    this.hud = null;
-
-    // Métricas de desempenho
-    this.acertosSemContaminacao = 0;
-    this.errosContaminacao = 0;
-    this.tempoLeituraRotulo = [];
-    this.labelTimerStart = null;
-  }
-
-  preload() {
-    const { phaseConfig } = this;
-    phaseConfig.items.forEach((item) => {
-      // Always load the main sprite for the item
-      this.load.image(item.key, item.spriteUrl);
-      // If the item defines a labelUrl property we load a separate image
-      // used for the zoomed label view. The key is suffixed with _label
-      // to avoid clashing with the regular sprite key. This enables a
-      // simple overlay when the player wants to inspect the product label.
-      if (item.labelUrl) {
-        this.load.image(`${item.key}_label`, item.labelUrl);
-      }
-    });
-    this.load.image('missing', '/assets/images/missing.png');
-
-    // HUD
-    this.load.image('uiPause', pauseIcon);
-    this.load.image('uiLifeFull', lifeFullIcon);
-    this.load.image('uiLifeEmpty', lifeEmptyIcon);
-    this.load.image('uiScore', scoreIcon);
-    this.load.image('uiTime', timeIcon);
-    this.load.image('uiTipBg', tipBgIcon);
-
-    // Chef sprites
-    this.load.image('chefIdle', '/assets/images/chef/idle.png');
-    this.load.image('chefCollect', '/assets/images/chef/collect.png');
-    this.load.image('chefMiss', '/assets/images/chef/miss.png');
-
-    if (phaseConfig.background) {
-      this.load.image('background', phaseConfig.background);
-    }
-
-    // Sons
-    this.load.audio('bgMusic', this.music);
-    this.load.audio('safeSound', '/assets/audio/safe.ogg');
-    this.load.audio('allergenSound', '/assets/audio/allergen.ogg');
-    this.load.audio('glutenSound', '/assets/audio/allergen.ogg'); // provisório
-  }
-
-  create() {
-    const { width, height } = this.scale;
-    const base = Math.max(height, 320);
-    const { phaseConfig, engine } = this;
-
-    const chefSize = Math.max(base * 0.15, 32);
-    this.hud = getHUDConfig(width, height, chefSize);
-    const { margin, iconSize, textStyle } = this.hud;
-
-    this.hudContainer = this.add.container(0, 0).setDepth(10);
-
-    if (phaseConfig.background) {
-      const bg = this.add.image(width / 2, height / 2, 'background');
-      const scaleX = width / bg.width;
-      const scaleY = height / bg.height;
-      const scale = Math.max(scaleX, scaleY);
-      bg.setScale(scale).setScrollFactor(0);
-    }
-
-    this.bgMusic = this.sound.add('bgMusic', { loop: true });
-    this.bgMusic.play();
-    if (engine) {
-      engine.bgMusic = this.bgMusic;
-    }
-
-    // HUD Score
-    this.scoreContainer = this.add.container(margin, margin);
-    this.scoreIcon = this.add.image(0, 0, 'uiScore').setOrigin(0, 0.5).setDisplaySize(iconSize, iconSize);
-    this.scoreText = this.add.text(iconSize + margin / 2, 0, `${this.score}`, textStyle).setOrigin(0, 0.5);
-    this.scoreContainer.add([this.scoreIcon, this.scoreText]);
-    this.hudContainer.add(this.scoreContainer);
-
-    // HUD Vidas
-    this.livesContainer = this.add.container(width - margin, margin);
-    for (let i = 0; i < this.lives; i++) {
-      const icon = this.add
-        .image(-i * (iconSize + margin / 2), 0, 'uiLifeFull')
-        .setOrigin(1, 0.5)
-        .setDisplaySize(iconSize, iconSize);
-      this.lifeIcons.push(icon);
-      this.livesContainer.add(icon);
-    }
-    this.hudContainer.add(this.livesContainer);
-
-    // HUD Tempo
-    if (this.duration) {
-      this.remainingTime = this.duration;
-      this.timeContainer = this.add.container(width / 2, margin);
-      this.timeIcon = this.add
-        .image(-iconSize - margin / 2, 0, 'uiTime')
-        .setOrigin(0, 0.5)
-        .setDisplaySize(iconSize, iconSize);
-      this.timeText = this.add.text(0, 0, `${this.remainingTime}`, textStyle).setOrigin(0, 0.5);
-      this.timeContainer.add([this.timeIcon, this.timeText]);
-      this.hudContainer.add(this.timeContainer);
-      this.timeEvent = this.time.addEvent({
-        delay: 1000,
-        callback: () => {
-          this.remainingTime -= 1;
-          this.timeText.setText(`${this.remainingTime}`);
-          this.animateHUD(this.timeContainer);
-        },
-        loop: true,
-      });
-    }
-
-    // HUD Dicas
-    this.tipCard = this.add
-      .image(width / 2, this.hud.tip.y, 'uiTipBg')
-      .setOrigin(0.5)
-      .setDepth(9)
-      .setVisible(false)
-      .setDisplaySize(this.hud.tip.cardWidth, this.hud.tip.cardHeight);
-    this.tipText = this.add
-      .text(width / 2, this.hud.tip.y, '', {
-        fontSize: `${iconSize * 0.6}px`,
-        fill: '#000',
-        align: 'center',
-        wordWrap: { width: this.hud.tip.cardWidth - margin * 2 },
-      })
-      .setOrigin(0.5)
-      .setDepth(10)
-      .setVisible(false);
-
-    // Chef
-    this.chef = this.add
-      .sprite(width / 2, height - chefSize / 2 - margin * 2, 'chefIdle')
-      .setDepth(5)
-      .setDisplaySize(chefSize, chefSize);
-
-    // Pause
-    this.pauseButton = this.add
-      .image(width - margin * 2, margin * 2, 'uiPause')
-      .setDisplaySize(iconSize, iconSize)
-      .setInteractive();
-    const launchPause = () => {
-      this.bgMusic?.pause();
-      this.scene.launch('PauseScene');
-      this.scene.pause();
-    };
-    this.pauseButton.on('pointerdown', launchPause);
-    this.input.keyboard.on('keydown-P', launchPause);
-
-    // Spawn loop
-    this.spawnLoop = this.time.addEvent({
-      delay: this.spawnRate,
-      loop: true,
-      callback: () => {
-        if (this.activeItems.length < this.simultaneous) {
-          this.spawnItem();
-        }
-      },
-    });
-
-    // Timer final de fase
-    if (this.duration) {
-      this.time.delayedCall(this.duration * 1000, () => {
-        this.endGame(true);
-      });
-    }
-
-    this.scale.on('resize', this.handleResize, this);
-    this.events.on('resize', this.handleResize, this);
-  }
-
-  handleResize(gameSize) {
-    const { width, height } = gameSize;
-    const base = Math.max(height, 320);
-    const chefSize = Math.max(base * 0.15, 32);
-
-    this.hud = getHUDConfig(width, height, chefSize);
-    const { margin, iconSize, textStyle } = this.hud;
-
-    if (this.scoreContainer) {
-      this.scoreContainer.setPosition(margin, margin);
-      this.scoreIcon.setDisplaySize(iconSize, iconSize);
-      this.scoreText
-        .setPosition(iconSize + margin / 2, 0)
-        .setStyle(textStyle);
-    }
-
-    if (this.livesContainer) {
-      this.livesContainer.setPosition(width - margin, margin);
-      this.lifeIcons.forEach((icon, index) => {
-        icon
-          .setDisplaySize(iconSize, iconSize)
-          .setPosition(-index * (iconSize + margin / 2), 0);
-      });
-    }
-
-    if (this.timeContainer) {
-      this.timeContainer.setPosition(width / 2, margin);
-      this.timeIcon
-        .setDisplaySize(iconSize, iconSize)
-        .setPosition(-iconSize - margin / 2, 0);
-      this.timeText.setPosition(0, 0).setStyle(textStyle);
-    }
-
-    if (this.tipCard && this.tipText) {
-      this.updateTipLayout();
-    }
-
-    if (this.chef) {
-      this.chef
-        .setPosition(width / 2, height - chefSize / 2 - margin * 2)
-        .setDisplaySize(chefSize, chefSize);
-    }
-
-    if (this.pauseButton) {
-      this.pauseButton
-        .setPosition(width - margin * 2, margin * 2)
-        .setDisplaySize(iconSize, iconSize);
-    }
-  }
-
-  updateTipLayout() {
-    if (!this.tipCard || !this.tipText || !this.hud) return;
-    const { width } = this.scale;
-    const { margin, iconSize, tip } = this.hud;
-    this.tipCard
-      .setDisplaySize(tip.cardWidth, tip.cardHeight)
-      .setPosition(width / 2, tip.y);
-    this.tipText
-      .setPosition(width / 2, tip.y)
-      .setStyle({
-        fontSize: `${iconSize * 0.6}px`,
-        fill: '#000',
-        align: 'center',
-        wordWrap: { width: tip.cardWidth - margin * 2 },
-      });
-  }
-
-  spawnItem() {
-    const { width, height } = this.scale;
-    const base = Math.max(height, 320);
-    const phaseConfig = this.phaseConfig;
-    // Randomly select an item definition from the phase configuration.
-    const item = Phaser.Utils.Array.GetRandom(phaseConfig.items);
-    const sizeRatio = phaseConfig.itemScale || 0.1;
-    const itemSize = phaseConfig.itemSize || Math.max(base * sizeRatio, 32);
-
-    let sprite;
-    // Special handling for runner/horizontal items. Runner items travel across
-    // the screen horizontally instead of falling vertically. They are
-    // configured with a `runner` flag and optional `direction` and `speed`.
-    if (item.runner) {
-      // Determine horizontal spawn side based on direction (default to left).
-      const dir = item.direction === 'right' ? 1 : -1;
-      // Randomly choose a vertical position within the screen for horizontal items.
-      const y = Phaser.Math.Between(itemSize / 2, height - itemSize / 2);
-      const x = dir > 0 ? -itemSize / 2 : width + itemSize / 2;
-      sprite = this.physics.add.image(x, y, item.key);
-      sprite.setDisplaySize(itemSize, itemSize);
-      // Save final scale values and start invisible for a pop-in animation.
-      const finalScaleX = sprite.scaleX;
-      const finalScaleY = sprite.scaleY;
-      sprite.setScale(0);
-      sprite.setData('itemData', item);
-      // Set horizontal velocity based on configured speed or default. No vertical velocity.
-      const horizontalSpeed = item.speed || 150;
-      sprite.setVelocityX(horizontalSpeed * dir);
-      sprite.setVelocityY(0);
-      // Animate the item scaling in using a tween.
-      this.tweens.add({
-        targets: sprite,
-        scaleX: { from: 0, to: finalScaleX },
-        scaleY: { from: 0, to: finalScaleY },
-        duration: 300,
-        ease: 'Back.Out',
-      });
-    } else {
-      // Default behaviour for vertically falling items.
-      const x = Phaser.Math.Between(itemSize / 2, width - itemSize / 2);
-      sprite = this.physics.add.image(x, -itemSize / 2, item.key);
-      sprite.setDisplaySize(itemSize, itemSize);
-      const finalScaleX = sprite.scaleX;
-      const finalScaleY = sprite.scaleY;
-      sprite.setScale(0);
-      sprite.setData('itemData', item);
-      // Apply vertical velocity based on phase speed.
-      sprite.setVelocityY(100 * this.speed);
-      // Pop-in animation.
-      this.tweens.add({
-        targets: sprite,
-        scaleX: { from: 0, to: finalScaleX },
-        scaleY: { from: 0, to: finalScaleY },
-        duration: 300,
-        ease: 'Back.Out',
-      });
-    }
-
-    // Items are interactive so players can click/tap them.
-    sprite.setInteractive();
-    sprite.on('pointerdown', () => this.handleItemClick(sprite));
-    this.activeItems.push(sprite);
-  }
-
-  handleItemClick(sprite) {
-    // Record how long the user spent reading the previous label (if any).
-    this.recordLabelTime();
-    const item = sprite.getData('itemData');
-    if (!item) return;
-
-    // Explosive items trigger an immediate penalty and shake effect. They are
-    // intentionally punishing and bypass normal gluten logic.
-    if (item.explosive) {
-      this.handleExplosive(sprite, item);
-      return;
-    }
-
-    // Items with a labelUrl require the player to read the label before
-    // classification occurs. We pause normal processing, display the label
-    // overlay and then continue classification after a short delay.
-    if (item.labelUrl) {
-      this.showLabel(item, sprite);
-      return;
-    }
-
-    // Standard items are processed immediately.
-    this.processItem(item, sprite);
-  }
-
-  /**
-   * Classify an item based on its properties and the player's allergen bitmask.
-   * Returns an outcome object describing the result. This helper centralises
-   * the logic used for all items (normal and after label inspection).
-   */
-  classifyItem(item) {
-    const bit = item.bitmaskBit;
-    const isAllergen = typeof bit === 'number' && (this.bitmask & (1 << bit)) !== 0;
-    const { containsGluten, crossContamination } = item;
-    if (containsGluten) {
-      return { type: 'gluten' };
-    }
-    if (crossContamination) {
-      return { type: 'contamination', penalty: item.penalty || 5 };
-    }
-    if (isAllergen) {
-      return { type: 'allergen' };
-    }
-    if (item.trap) {
-      return { type: 'trap', penalty: item.penalty || 10 };
-    }
-    return { type: 'correct', bonus: item.bonus || 0 };
-  }
-
-  /**
-   * Processes an item click after any optional label reading. Handles
-   * classification, scoring, life updates and removal. All normal item
-   * interactions ultimately flow through this helper.
-   */
-  processItem(item, sprite) {
-    const outcome = this.classifyItem(item);
-    // Determine whether the player was correct based on outcome
-    const correct = outcome.type === 'correct';
-    switch (outcome.type) {
-      case 'gluten':
-        this.animateChef('miss');
-        this.sound.play('glutenSound');
-        this.lives -= 1;
-        this.updateLives();
-        this.errosContaminacao += 1;
-        this.showTip();
-        break;
-      case 'contamination':
-        this.animateChef('miss');
-        this.sound.play('allergenSound');
-        this.score = Math.max(0, this.score - outcome.penalty);
-        this.updateScore();
-        this.errosContaminacao += 1;
-        this.showTip();
-        break;
-      case 'allergen':
-        this.animateChef('miss');
-        this.sound.play('allergenSound');
-        this.lives -= 1;
-        this.updateLives();
-        this.showTip();
-        break;
-      case 'trap':
-        this.animateChef('miss');
-        this.sound.play('allergenSound');
-        this.score = Math.max(0, this.score - outcome.penalty);
-        this.updateScore();
-        break;
-      case 'correct':
-      default:
-        this.animateChef('collect');
-        this.sound.play('safeSound');
-        this.score += 10 + (outcome.bonus || 0);
-        this.updateScore();
-        this.acertosSemContaminacao += 1;
-        break;
-    }
-    // Update difficulty based on correctness
-    this.difficulty.record(correct);
-    const params = this.difficulty.getParameters(this.baseSpawnRate, this.baseSimultaneous);
-    this.spawnRate = params.spawnRate;
-    this.simultaneous = params.simultaneous;
-    if (this.spawnLoop) this.spawnLoop.delay = this.spawnRate;
-
-    // Safely remove the sprite with a fade/scale animation
-    sprite.disableInteractive();
-    if (sprite.body) sprite.body.enable = false;
-    this.activeItems = this.activeItems.filter((i) => i !== sprite);
-    this.tweens.add({
-      targets: sprite,
-      scale: 0,
-      alpha: 0,
-      duration: 200,
-      onComplete: () => {
-        sprite.destroy();
-        if (this.lives <= 0) {
-          this.endGame();
-        }
-      },
-    });
-  }
-
-  /**
-   * Handles explosive items. These objects cause a bigger penalty and a
-   * visual effect. Explosive items ignore all other classification logic.
-   */
-  handleExplosive(sprite, item) {
-    // Penalise two lives by default when exploding
-    this.animateChef('miss');
-    this.sound.play('allergenSound');
-    const penaltyLives = item.penaltyLives || 2;
-    const penaltyScore = item.penalty || 15;
-    this.lives = Math.max(0, this.lives - penaltyLives);
-    this.updateLives();
-    this.score = Math.max(0, this.score - penaltyScore);
-    this.updateScore();
-    // Shake the camera to signal an explosion
-    if (this.cameras && this.cameras.main) {
-      this.cameras.main.shake(300, 0.03);
-    }
-    // Remove the explosive sprite after the effect
-    sprite.disableInteractive();
-    if (sprite.body) sprite.body.enable = false;
-    this.activeItems = this.activeItems.filter((i) => i !== sprite);
-    this.tweens.add({
-      targets: sprite,
-      scale: 0,
-      alpha: 0,
-      duration: 200,
-      onComplete: () => {
-        sprite.destroy();
-        if (this.lives <= 0) {
-          this.endGame();
-        }
-      },
-    });
-  }
-
-  /**
-   * Displays an overlay showing the product label image for items that
-   * define a `labelUrl`. The game briefly pauses classification while
-   * the player inspects the label. After a short delay the item is
-   * processed as usual via processItem().
-   */
-  showLabel(item, sprite) {
-    // Draw semi-transparent backdrop
-    const overlay = this.add.rectangle(
-      this.scale.width / 2,
-      this.scale.height / 2,
-      this.scale.width,
-      this.scale.height,
-      0x000000,
-      0.7
-    ).setDepth(50);
-    // Determine the key for the label image (loaded during preload)
-    const labelKey = `${item.key}_label`;
-    const labelImage = this.add.image(
-      this.scale.width / 2,
-      this.scale.height / 2,
-      labelKey
-    ).setDepth(51);
-    // Scale the label to fit within 80% of the viewport
-    const scaleX = (this.scale.width * 0.8) / labelImage.width;
-    const scaleY = (this.scale.height * 0.8) / labelImage.height;
-    const scale = Math.min(scaleX, scaleY);
-    labelImage.setScale(scale);
-    // After a brief period, remove overlay and process the item normally
-    this.time.delayedCall(2500, () => {
-      overlay.destroy();
-      labelImage.destroy();
-      this.processItem(item, sprite);
-    });
-  }
-
-  updateScore() {
-    this.scoreText.setText(`${this.score}`);
-    this.animateHUD(this.scoreContainer);
-  }
-
-  updateLives() {
-    this.lifeIcons.forEach((icon, index) => {
-      const key = index < this.lives ? 'uiLifeFull' : 'uiLifeEmpty';
-      icon.setTexture(key);
-    });
-    this.animateHUD(this.livesContainer);
-  }
-
-  showTip() {
-    if (this.tips.length === 0) return;
-    this.hud = getHUDConfig(this.scale.width, this.scale.height, this.chef.displayHeight);
-    this.updateTipLayout();
-    const tip = Phaser.Utils.Array.GetRandom(this.tips);
-    this.tipCard.setVisible(true);
-    this.tipText.setText(tip).setVisible(true);
-    if (this.tipTimer) this.tipTimer.remove(false);
-    this.tipTimer = this.time.addEvent({
-      delay: 5000,
-      callback: () => {
-        this.tipText.setText('').setVisible(false);
-        this.tipCard.setVisible(false);
-        this.tipTimer = null;
-      },
-    });
-  }
-
-  recordLabelTime() {
-    if (this.labelTimerStart != null) {
-      const elapsed = (this.time.now - this.labelTimerStart) / 1000;
-      this.tempoLeituraRotulo.push(elapsed);
-      this.labelTimerStart = null;
-    }
-  }
-
-  endGame(success = false) {
-    this.scene.pause();
-    this.bgMusic.stop();
-    this.add
-      .text(this.scale.width / 2, this.scale.height / 2, 'Fim de Jogo', {
-        fontSize: '32px',
-        fill: '#f00',
-      })
-      .setOrigin(0.5);
-    if (typeof this.onGameOver === 'function') {
-      this.onGameOver({
-        success,
-        acertosSemContaminacao: this.acertosSemContaminacao,
-        errosContaminacao: this.errosContaminacao,
-        tempoLeituraRotulo: this.tempoLeituraRotulo,
-      });
-    }
-  }
-
-  update() {
-    // Remove any items that have moved completely off screen. This applies
-    // both to vertical items that fall past the bottom edge and horizontal
-    // runner items that travel beyond the left/right edges. Off-screen
-    // sprites are faded out and destroyed to free resources.
-    this.activeItems.forEach((sprite) => {
-      const offBottom = sprite.y > this.scale.height + sprite.displayHeight / 2;
-      const offLeft = sprite.x < -sprite.displayWidth / 2;
-      const offRight = sprite.x > this.scale.width + sprite.displayWidth / 2;
-      if (offBottom || offLeft || offRight) {
-        this.activeItems = this.activeItems.filter((i) => i !== sprite);
-        this.tweens.add({
-          targets: sprite,
-          alpha: 0,
-          duration: 200,
-          onComplete: () => sprite.destroy(),
-        });
-      }
-    });
-  }
-
-  animateChef(state) {
-    if (!this.chef) return;
-    const textures = {
-      idle: 'chefIdle',
-      collect: 'chefCollect',
-      miss: 'chefMiss',
-    };
-    this.chef.setTexture(textures[state] || textures.idle);
-    if (state !== 'idle') {
-      this.time.delayedCall(500, () => {
-        this.chef.setTexture('chefIdle');
-      });
-    }
-  }
-
-  animateHUD(target) {
-    this.tweens.add({
-      targets: target,
-      scale: { from: 1, to: 1.2 },
-      duration: 200,
-      yoyo: true,
-    });
-  }
-}
