--- feira.json
@@ -1,56 +1,0 @@
-{
-  "transition": {
-    "text": "Bem-vindo à Feira!",
-    "image": "/assets/images/bg/feira_day.png",
-    "tip": "Dica: alguns produtos na feira podem conter glúten escondido!",
-    "audio": "/assets/audio/safe.ogg",
-    "duration": 4000,
-    "skipAfter": 2000,
-    "animation": "slide",
-    "instructions": "Na Feira, colete frutas e legumes seguros. Evite itens com glúten como pães, trigo e produtos que possam estar contaminados. Fique atento aos ícones explosivos e itens bônus passando!"
-  },
-  "background": "/assets/images/bg/feira_day.png",
-  "music": "/assets/audio/background.ogg",
-  "speed": 0.9,
-  "itemSpawnRate": 1600,
-  "duration": 60,
-  "simultaneous": 1,
-  "itemScale": 0.07,
-  "items": [
-    {
-      "key": "apple",
-      "spriteUrl": "/assets/images/items/apple.png"
-    },
-    {
-      "key": "banana",
-      "spriteUrl": "/assets/images/items/banana.png",
-      "trap": true,
-      "penalty": 5,
-      "crossContamination": true
-    },
-    {
-      "key": "bread",
-      "spriteUrl": "/assets/images/items/bread.png",
-      "containsGluten": true,
-      "labelUrl": "/assets/images/items/bread.png"
-    },
-    {
-      "key": "wheat",
-      "spriteUrl": "/assets/images/items/wheat.png",
-      "containsGluten": true
-    },
-    {
-      "key": "pineapple",
-      "spriteUrl": "/assets/images/items/pineapple.png"
-    },
-    {
-      "key": "strawberry",
-      "spriteUrl": "/assets/images/items/strawberry.png",
-      "crossContamination": true
-    }
-  ],
-  "tips": [
-    "Pães e farinhas expostas carregam muito glúten.",
-    "Frutas cortadas com utensílios usados em massas podem contaminar."
-  ]
-}
