--- festa.json
@@ -1,59 +1,0 @@
-{
-  "transition": {
-    "text": "Festa de Aniversário chegando!",
-    "image": "/assets/images/bg/festa.png",
-    "animation": "fade",
-    "audio": "/assets/audio/safe.ogg",
-    "duration": 3000,
-    "skipAfter": 2000,
-    "tip": "Dica: doces podem esconder glúten nas receitas!",
-    "instructions": "Na Festa de Aniversário, muitos doces e bolos contêm glúten. Clique somente em itens seguros e fique atento a itens que se movem rapidamente, pois podem ser bônus!"
-  },
-  "background": "/assets/images/bg/festa.png",
-  "music": "/assets/audio/background.ogg",
-  "speed": 1.4,
-  "itemSpawnRate": 900,
-  "duration": 80,
-  "simultaneous": 3,
-  "itemScale": 0.07,
-  "items": [
-    {
-      "key": "cake",
-      "spriteUrl": "/assets/images/items/cake.png",
-      "containsGluten": true,
-      "labelUrl": "/assets/images/items/cake.png"
-    },
-    {
-      "key": "icecream",
-      "spriteUrl": "/assets/images/items/icecream.png",
-      "crossContamination": true
-    },
-    {
-      "key": "cookie",
-      "spriteUrl": "/assets/images/items/cookie.png",
-      "containsGluten": true
-    },
-    {
-      "key": "candy",
-      "spriteUrl": "/assets/images/items/candy.png",
-      "runner": true,
-      "direction": "right",
-      "speed": 200,
-      "bonus": 5
-    },
-    {
-      "key": "chocolate",
-      "spriteUrl": "/assets/images/items/chocolate.png",
-      "crossContamination": true
-    },
-    {
-      "key": "strawberry",
-      "spriteUrl": "/assets/images/items/strawberry.png",
-      "bonus": 20
-    }
-  ],
-  "tips": [
-    "Bolos e biscoitos são ricos em glúten.",
-    "Utilize utensílios limpos para evitar contaminação cruzada."
-  ]
-}
