--- praia.json
@@ -1,57 +1,0 @@
-{
-  "phaseId": "praia",
-  "transition": {
-    "text": "Vamos relaxar na Praia!",
-    "image": "/assets/images/bg/praia.png",
-    "animation": "fade",
-    "audio": "/assets/audio/safe.ogg",
-    "duration": 3000,
-    "skipAfter": 2000,
-    "tip": "Dica: empanados de praia podem conter farinha!",
-    "instructions": "Na Praia, muita comida frita é empanada com farinha de trigo. Clique apenas em frutas frescas e bebidas seguras. Alguns itens podem mover-se rapidamente trazendo bônus ou armadilhas."
-  },
-  "background": "/assets/images/bg/praia.png",
-  "music": "/assets/audio/beach_loop.mp3",
-  "speed": 1.6,
-  "itemSpawnRate": 700,
-  "duration": 90,
-  "simultaneous": 4,
-  "itemScale": 0.07,
-  "items": [
-    {
-      "key": "shrimp",
-      "spriteUrl": "/assets/images/items/shrimp.png",
-      "crossContamination": true
-    },
-    {
-      "key": "fish",
-      "spriteUrl": "/assets/images/items/fish.png",
-      "crossContamination": true
-    },
-    {
-      "key": "squid",
-      "spriteUrl": "/assets/images/items/squid.png",
-      "crossContamination": true,
-      "trap": true,
-      "penalty": 5
-    },
-    {
-      "key": "cake",
-      "spriteUrl": "/assets/images/items/cake.png",
-      "containsGluten": true
-    },
-    {
-      "key": "coconut",
-      "spriteUrl": "/assets/images/items/coconut.png"
-    },
-    {
-      "key": "pineapple",
-      "spriteUrl": "/assets/images/items/pineapple.png",
-      "bonus": 15
-    }
-  ],
-  "tips": [
-    "Frituras empanadas costumam usar farinha de trigo.",
-    "Separe petiscos 'sem glúten' de outros utensílios."
-  ]
-}
