--- supermercado.json
@@ -1,67 +1,0 @@
-{
-  "transition": {
-    "text": "Hora das compras no Supermercado!",
-    "image": "/assets/images/bg/supermercado.png",
-    "animation": "fade",
-    "audio": "/assets/audio/safe.ogg",
-    "duration": 3000,
-    "skipAfter": 2000,
-    "tip": "Dica: procure rótulos 'sem glúten' nos produtos!",
-    "instructions": "No Supermercado, leia atentamente os rótulos e clique apenas em produtos sem glúten. Itens com selo 'sem glúten' são seguros. Tenha cuidado com contaminação cruzada e com itens explosivos!"
-  },
-  "background": "/assets/images/bg/supermercado.png",
-  "music": "/assets/audio/background.ogg",
-  "speed": 1.2,
-  "itemSpawnRate": 1300,
-  "duration": 70,
-  "simultaneous": 2,
-  "itemScale": 0.07,
-  "items": [
-    {
-      "key": "cookie",
-      "spriteUrl": "/assets/images/items/cookie.png",
-      "containsGluten": true
-    },
-    {
-      "key": "milk",
-      "spriteUrl": "/assets/images/items/milk.png"
-    },
-    {
-      "key": "bread",
-      "spriteUrl": "/assets/images/items/bread.png",
-      "containsGluten": true,
-      "trap": true,
-      "penalty": 10
-    },
-    {
-      "key": "cereal",
-      "spriteUrl": "/assets/images/items/cereal.png",
-      "containsGluten": true
-    },
-    {
-      "key": "cheese",
-      "spriteUrl": "/assets/images/items/cheese.png",
-      "crossContamination": true
-    },
-    {
-      "key": "egg",
-      "spriteUrl": "/assets/images/items/egg.png"
-    },
-    {
-      "key": "pineapple",
-      "spriteUrl": "/assets/images/items/pineapple.png",
-      "bonus": 15
-    }
-    ,
-    {
-      "key": "hazelnut",
-      "spriteUrl": "/assets/images/items/hazelnut.png",
-      "explosive": true,
-      "penalty": 15
-    }
-  ],
-  "tips": [
-    "Leia os rótulos e procure por produtos com selo 'sem glúten'.",
-    "Itens a granel podem ter contato com farinhas."
-  ]
-}
